Fix file imports and remove unused module
