+++ client/src/app/metadata-schema-form/metadata-schema-form.module.ts
@@ -80,7 +80,7 @@ import { VfTabComponent } from './custom/vf-tab/vf-tab.component';
     BaseInputComponent,
     SelectComponent,
     EnumDropDownComponent,
-    SelectListComponent
+    SelectListComponent,
   ],
   providers: []
 })
