--- client/src/app/project-registration/contact-name-field/contact-name-field.component.html
@@ -1,19 +1,0 @@
-<app-vf-input
-  [isRequired]="true"
-  [label]="'First Name'"
-  [dataType]="'string'"
-  [error]="control.touched && firstNameCtrl.errors && firstNameCtrl.errors.required ? 'This field is required': ''"
-  (blur)="onBlur()"
-  [formControl]="firstNameCtrl"
->
-</app-vf-input>
-
-<app-vf-input
-  [isRequired]="true"
-  [label]="'Last Name'"
-  [dataType]="'string'"
-  [error]="control.touched && lastNameCtrl.errors && lastNameCtrl.errors.required ? 'This field is required': ''"
-  [formControl]="lastNameCtrl"
-  (blur)="onBlur()"
->
-</app-vf-input>
