--- client/src/app/project-registration/project-registration-form/project-registration-form.component.html
@@ -1,31 +1,0 @@
-<h1 class="vf-text vf-text-heading--2">{{title}}</h1>
-<br/>
-<br/>
-
-<mat-tab-group>
-  <mat-tab label="1. Project">
-    <ng-container *ngIf= "(userAccount$ | async) && projectFormData">
-      <div class="vf-u-margin__top--xxl">&nbsp;</div>
-
-      <app-metadata-form name="project"
-                         [useVf]="true"
-                         [schema]="projectIngestSchema"
-                         [data]="projectFormData"
-                         [config]="config"
-                         [selectedTabKey]="getCurrentTab()"
-                         (cancel)="onCancel($event)"
-                         (back)="decrementProjectTab()"
-                         (tabChange)="onTabChange($event)"
-                         (save)="onSave($event)"
-                         (formValueChange)="saveProjectInCache($event)">
-      </app-metadata-form>
-    </ng-container>
-  </mat-tab>
-  <mat-tab label="2. Experiment Information" disabled></mat-tab>
-  <mat-tab label="3. Data upload" disabled></mat-tab>
-
-  <ng-container *ngIf="userIsWrangler">
-    <mat-tab label="4. View Metadata" disabled></mat-tab>
-    <mat-tab label="5. View Data" disabled></mat-tab>
-  </ng-container>
-</mat-tab-group>
