--- client/src/app/project-registration/wrangler-list-input/wrangler-list-input.component.html
@@ -1,14 +1,0 @@
-<app-select *ngIf="!control.value || selectedWrangler"
-  [id]="id"
-  [label]="label"
-  [isRequired]="isRequired"
-  [disabled]="disabled"
-  [placeholder]="placeholder"
-  [helperText]="helperText"
-  [options$]="options$"
-  [displayWith]="displayWranglerName"
-  [value]="selectedWrangler"
-  [error]="checkForErrors(control)"
-  (searchValueChanged)="searchControl.setValue($event);"
-  (selectedValueChanged)="updateControl($event)"
-></app-select>
