--- client/src/app/project-registration/wrangling-priority-input/wrangling-priority-input.component.html
@@ -1,13 +1,0 @@
-<app-select-list
-  [id]="id"
-  [label]="label"
-  [isRequired]="isRequired"
-  [disabled]="disabled"
-  [placeholder]="placeholder"
-  [helperText]="helperText"
-  [options]="options"
-  [removeBlankOption]=true
-  [value]="valueOf(control.value)"
-  [error]="checkForErrors(control)"
-  (valueChanged)="updateControl($event)"
-></app-select-list>
